+++ test/__init__.py
@@ -1,4 +1,14 @@
 import os
+import shutil
 
-if not os.path.exists("tmp/"): 
-    os.makedirs("tmp/") 
+
+if os.path.exists("dist/"):
+    shutil.rmtree("dist/")
+
+if os.path.exists("pycaptions.egg-info/"):
+    shutil.rmtree("pycaptions.egg-info/")
+
+if os.path.exists("tmp/"): 
+    shutil.rmtree("tmp/")
+
+os.makedirs("tmp/")